se termino el modulo de acompanante

+++ src/main/java/com/maryed/hotel_admin_system/cliente/service/ClienteServiceImpl.java
@@ -83,6 +83,9 @@ public class ClienteServiceImpl implements ClienteService {
 
     @Override
     public void eliminar(Integer id) {
+        if (!repository.existsById(id)) {
+            throw new ClienteServiceException("cliente no encontrado para eliminar");
+        }
         repository.deleteById(id);
     }
 }
